Engine: Fixed bug in `run_app` function
We were calling `cbegin()` to get a pointer to the underlying data,
however, this actually returns an iterator. MSVC complains about
receiving an iterator when it expected a pointer. We now use `data()`,
which returns a pointer to the cstring.

Also fixed a bug in `main.cpp` where we were passing an int to the
Circle constructor that expects a float.