Engine: Greatly improved the main loop
Previously, we would try and render every iteration of the main loop.
This is likely to lead to very inconsistent behaviour between different
systems.

The idea here, is to separate update logic from render logic, and only
render at most 60 times per seconds. We are basically capping the frame
time at 12ms, but still updating the game at the same rate internally.

This will allow us to not waste time rendering the scene excessively,
and focus on running the physics simulation